Simplify config command and processing
While corax is still in development and not quite matured, the config
command is going to remain simple and give limited functionality
for ways in which to configure that app. As such, the option
parameters have been removed for now. This update also introduces the
config.yaml file which serves as an example and template on first
configuration run.