Rename reputation command to analyze
The working being performed by what was the reputation command
seemed more like a command to analyze a given observable. As such,
reputation has been changed to analyze. Additionally, the reading
of the configuration file has been moved to the functions which
require the configurations. While this results in code not being DRY,
it is a quick fix to the logic bug it introduced.